requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# YOLO
ultralytics>=8.0.0

# SAM
segment-anything>=1.0
git+https://github.com/facebookresearch/segment-anything.git

# Evaluation metrics
scikit-learn>=1.3.0
pycocotools>=2.0.6

# Visualization
plotly>=5.14.0